requirements.txt
selenium 
beautifulsoup4
webdriver-manager
pdfplumber
pymupdf
openai
chromadb
pandas
tiktoken
sentence-transformers
fuzzywuzzy
streamlit
openai
chromadb
sentence-transformers
langchain
langchain_community
python-dotenv

